requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
Flask-DB
pymysql
flask-jwt-extended
passlib